requirements.txt
streamlit
matplotlib
pandas
Pillow
koreanize-matplotlib



